// src/common/enum/order.enum.ts
export enum PaymentMethod {
    Cash = "cash",
    Card = "card"
}

export enum PaymentStatus {
    Paid = "paid",
    Unpaid = "unpaid",
    Pending = "pending",
    Refunded = "refunded",
}

export enum OrderStatus {
    Pending = "pending",
    Shipped = "shipped",
    Delivered = "delivered",
    Cancelled = "cancelled"
}